parser: add reservation map iterator
The reservation map iterator has been implemented as a macro instead of
taking a function pointer so that it can be easier to iterate and use
external variables. This way, if outside the iterator there is a
variable that needs to be accessed we don't need to pass it around as
void* and we can simply access it.