parser: implement dtb_find for other nodes
The parser basically goes through all the tokens and compares the name
of the node the token is representing with the part of the "device
path".

Important to note that when searching the '/memory' node, the
'/memory@800000' node from the device tree will be returned as the find
function basically ignores the unit name (what comes after the @).