test: add another skip_until_same_depth test
The test used before basically was designed to test if the function
correctly discarded subnodes. For example,
skip_until_same_depth("/cpus") returns the /memory node (skipping the
/cpus/cpu@0 and /cpus/cpu@1 subnodes).

The new test is designed to see if the function behaves correctly when
there are no subnodes, in this case skip_until_same_depth("/cpus/cpu@0")
should return the /cpus/cpu@1 node.