test: unify test printing and make it pretier
There is now a function called print_test_result that takes a condition
and prints if either the test succeded or failed. The function also
aborts now instead of exiting so that debuggers can be easilly used.

To make it pretier there is MAX_LINE_WIDTH define that defines how long
the line is going to be and the SUCCESS/FAILED text is right aligned.