parser: fix BEGIN_NODE finding on skip_until_same_depth
After an END_NODE token only END or NOP tokens can be found. Before the
logic tried (and failed) by simply advancing the tokens until either an
END node or a BEGIN node was found.

Now it simply advances on all the NOPs and checks if an END token has
been reached.