parser: proper address-cell and size-cell handling
Previously address-cell and size-cell properties where treated as normal
properties. Instead, now they are parsed by the iterator and saved as
variables to be used by the iterator user.

This commit assumes that both the address-cell and size-cell properties
come before any other property. All device trees I have seen do this,
but its not really on the spec so this should be looked at in the
future.