parser: skip padding only if there on DTB_BEGIN_NODE
DTB_BEGIN_NODE is followed by a null terminated string representing the
name of the node. After that there are 0s to align to the 32 bit
boundary and some times another token of 0.

The memory node did not have an extra 0 of 32 bit padding, but the
parser assumed that there was always going to be that padding so it
skipped a node. That's now fixed.